requirements.txt
bleach==3.3.0
certifi==2020.6.20
chardet==3.0.4
colorama==0.4.3
django-widget-tweaks==1.4.8
docutils==0.16
idna==2.10
importlib-metadata==4.8.1
keyring==21.4.0
packaging==20.4
pkginfo==1.5.0.1
Pygments==2.7.4
pyparsing==2.4.7
readme-renderer==26.0
requests==2.24.0
requests-toolbelt==0.9.1
rfc3986==1.4.0
rstcheck==3.3.1
six==1.15.0
tqdm==4.48.2
twine==3.4.2
urllib3==1.25.11
webencodings==0.5.1
zipp==3.1.0
